Fix recreating APN list after other SIM is removed
The dummy SUB ID can be returned while the SUB to slot map is
rebuilding due to race condition, so that the previous SUB is
written as dummy SUB ID even its SIM is still present.

To fix it, avoid writing the previous SUB cache while SUB is not
active, after SIM is not ready, the previous SUB cache can be reset.

Change-Id: I08ed7706a4adc8943783c53c879cc6f9af17a0bb
CRs-Fixed: 2284030
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>